+++ client/src/components/BulmaAdvancedTable.vue
@@ -232,6 +232,7 @@
 
       },
       select(i){
+        console.log("A, we selecteren " + i)
         if(this.multiple){
           Vue.set(this.selected,i,!this.selected[i]);
         }else{
@@ -320,16 +321,17 @@
           }
           if(this.defaultValue=="__auto__" && this.values.length>0){
             this.select(0) // if __auto__ select the first
-          }else if(this.defaultValue=="__all__"){ // if a regular default is set, we select it
+          }else if(this.defaultValue=="__all__" && this.multiple){ // if all is set, we select all
             this.values.forEach((item,i) => {
               this.select(i)
             })
           }else if(this.defaultValue!="__none__" && this.defaultValue!=undefined){ // if a regular default is set, we select it
             var obj=undefined
             var defaulttype
-            try{
+             try{
               obj = JSON.parse(this.defaultValue)
               if(typeof obj == "object"){
+                console.log("moh")
                 defaulttype="object"
               }
             }catch(e){
@@ -339,16 +341,15 @@
               obj=this.defaultValue
               defaulttype="object"
             }
-            if(defaulttype=="object"){
+            if(defaulttype=="object" && !Array.isArray(this.defaultValue)){
               // we search for the value by property
-
               if(obj){
                 // loop all values
                 this.values.forEach((item,i) => {
                   try{
 
                     if(this.objectEqual(obj,item)){
-                        this.select(i)
+                        ref.select(i)
                     }
                   }catch(e){
                     console.log("Bad defaultvalue : "  + e)
@@ -359,13 +360,11 @@
             }else{
               // we search for the value by string
               this.values.forEach((item,i) => {
-
                   if(item && ref.defaultValue==(item[ref.valueLabel]||item)){
-                    this.select(i)
-                  }
-                  if(ref.multiple && Array.isArray(ref.defaultValue)){
-                    if(item && ref.default.includes(item[ref.valueLabel]||item||false)){
-                      this.select(i)
+                    ref.select(i)
+                  }else if(ref.multiple && Array.isArray(ref.defaultValue) && ref.defaultValue.length>0){
+                    if(item && ref.defaultValue.includes(item[ref.valueLabel]||item||false)){
+                      ref.select(i)
                     }
                   }
               })