+++ client/src/components/Form.vue
@@ -107,7 +107,7 @@
                           <span
                             class="icon is-clickable has-text-success"
                             @click="setExpressionFieldViewable(field.name,true)"
-                            v-if="['expression','query','table'].includes(field.type) && !fieldOptions[field.name].viewable && !fieldOptions[field.name].editable"
+                            v-if="field.expression && !fieldOptions[field.name].viewable && !fieldOptions[field.name].editable"
                           >
                             <font-awesome-icon icon="eye" />
                           </span>
@@ -115,7 +115,7 @@
                           <span
                             class="icon is-clickable has-text-info"
                             @click="clip((field.type=='expression')?$v.form[field.name].$model:queryresults[field.name])"
-                            v-if="['expression','query','table'].includes(field.type) && fieldOptions[field.name].viewable && !fieldOptions[field.name].editable"
+                            v-if="field.expression && fieldOptions[field.name].viewable && !fieldOptions[field.name].editable"
                           >
                             <font-awesome-icon icon="copy" />
                           </span>
@@ -123,7 +123,7 @@
                           <span
                             class="icon is-clickable has-text-danger"
                             @click="setExpressionFieldViewable(field.name,false)"
-                            v-if="['expression','query','table'].includes(field.type) && fieldOptions[field.name].viewable && !fieldOptions[field.name].editable"
+                            v-if="field.expression && fieldOptions[field.name].viewable && !fieldOptions[field.name].editable"
                           >
                             <font-awesome-icon icon="times" />
                           </span>
@@ -152,8 +152,8 @@
                       <div v-if="field.type=='checkbox'">
                         <BulmaCheckRadio checktype="checkbox" v-model="$v.form[field.name].$model" :name="field.name" :type="{'is-danger is-block':$v.form[field.name].$invalid}" :label="field.placeholder" @change="evaluateDynamicFields(field.name)" />
                       </div>
-                      <!-- type = query -->
-                      <div v-if="field.type=='query'">
+                      <!-- type = enum/query -->
+                      <div v-if="field.type=='query' || field.type=='enum'">
                         <BulmaAdvancedSelect
                           v-show="!fieldOptions[field.name].viewable"
                           :defaultValue="defaults[field.name]"
@@ -161,9 +161,9 @@
                           :multiple="field.multiple||false"
                           :name="field.name"
                           :placeholder="field.placeholder||'Select...'"
-                          :values="queryresults[field.name]||[]"
+                          :values="field.values||queryresults[field.name]||[]"
                           :hasError="$v.form[field.name].$invalid"
-                          :isLoading="!['fixed','variable'].includes(dynamicFieldStatus[field.name])"
+                          :isLoading="!field.values && !['fixed','variable'].includes(dynamicFieldStatus[field.name])"
                           v-model="$v.form[field.name].$model"
                           :icon="field.icon"
                           :columns="field.columns||[]"
@@ -217,7 +217,7 @@
                         </div>
                       </div>
                       <!-- fields with icons -->
-                      <div :class="{'has-icons-left':!!field.icon && field.type!='query'}" class="control">
+                      <div :class="{'has-icons-left':!!field.icon && !['query','enum'].includes(field.type)}" class="control">
                         <!-- type = expression -->
                         <div v-if="field.type=='expression'" :class="{'is-loading':(dynamicFieldStatus[field.name]==undefined || dynamicFieldStatus[field.name]=='running') &! fieldOptions[field.name].editable}" class="control">
                           <div v-if="!fieldOptions[field.name].viewable">
@@ -265,33 +265,8 @@
                           type="number"
                           :placeholder="field.placeholder"
                           @change="evaluateDynamicFields(field.name)">
-                        <!-- type = enum -->
-                        <div v-if="field.type=='enum' && !field.multiple" class="select">
-                          <select
-                            :name="field.name"
-                            @focus="inputFocus"
-                            :class="{'is-danger':$v.form[field.name].$invalid}"
-                            v-model="$v.form[field.name].$model"
-                            @change="evaluateDynamicFields(field.name)">
-                            <option v-if="!field.required" value=""></option>
-                            <option v-for="option in field.values" :key="option" :selected="defaults[field.name]==option" :value="option">{{ option }}</option>
-                          </select>
-                        </div>
-                        <!-- type = multiple enum -->
-                        <div v-if="field.type=='enum' && (field.multiple||false)==true" class="select is-multiple">
-                          <select
-                            :name="field.name"
-                            @focus="inputFocus"
-                            :class="{'is-danger':$v.form[field.name].$invalid}"
-                            v-model="$v.form[field.name].$model"
-                            multiple
-                            :size="field.size">
-                            <option v-if="!field.required" value=""></option>
-                            <option v-for="option in field.values" :key="option" :selected="defaults[field.name].includes(option)" :value="option">{{ option }}</option>
-                          </select>
-                        </div>
                         <!-- add left icon, but not for query, because that's a component with icon builtin -->
-                        <span v-if="!!field.icon && field.type!='query' && !(field.type=='expression' && fieldOptions[field.name].viewable)" class="icon is-small is-left">
+                        <span v-if="!!field.icon && !['query','enum'].includes(field.type) && !(field.type=='expression' && fieldOptions[field.name].viewable)" class="icon is-small is-left">
                           <font-awesome-icon :icon="field.icon" />
                         </span>
                       </div>
@@ -468,13 +443,13 @@
         var regexObj
         var description
         // required validation for simple fields
-        if(ff.type!='checkbox' && ff.type!='expression' && ff.type!='query'){
+        if(ff.type!='checkbox' && ff.type!='expression' && ff.type!='query' && ff.type!='enum'){
           attrs.required=requiredIf(function(){
             return !!ff.required
           })
         }
         // required validation for expression field
-        if(((ff.type=="expression")||(ff.type=="query")) && ff.required){
+        if(((ff.type=="expression")||(ff.type=="query")||(ff.type=="enum")) && ff.required){
           attrs.required=helpers.withParams(
               {description: "This field is required"},
               (value) => (value!=undefined && value!=null && value!='__auto__' && value!='__none__' && value!='__all__')
@@ -879,7 +854,13 @@
           if(item.sameAs && !fields.includes(item.sameAs)){
             ref.warnings.push(`<span class="has-text-warning-dark">'${item.name}' has bad 'sameAs' validation</span><br><span>${item.sameAs} is not a valid field name</span>`)
           }
-          if(["expression","query","table"].includes(item.type)){
+          if(["expression","query","enum","table"].includes(item.type)){
+
+            // query type is now deprecated
+            if(item.type=='query'){
+              ref.warnings.push(`<span class="has-text-warning-dark">'${item.name}' has the deprecated query type</span><br><span>Use enum type instead.</span>`)            
+            }
+            
             var testRegex = /\$\(([^)]+)\)/g;
             var matches=(item.expression || item.query || "").matchAll(testRegex);
             for(var match of matches){
@@ -895,7 +876,7 @@
               }
               foundfield=foundfield.replace(/\[[0-9]*\]/,'') // xxx[y] => xxx
               if(fields.includes(foundfield)){                         // does field xxx exist in our form ?
-                // console.log(foundfield + " is a real field")
+                //console.log(foundfield + " is a real field")
                 if(foundfield in ref.dynamicFieldDependencies){															 // did we declare it before ?
                   if(ref.dynamicFieldDependencies[foundfield].indexOf(item.name) === -1) {  // allready in there ?
                       ref.dynamicFieldDependencies[foundfield].push(item.name);												 // push it
@@ -1243,7 +1224,7 @@
                         try{
                           result=eval(placeholderCheck.value)
                           if(item.type=="expression") Vue.set(ref.form, item.name, result);
-                          if(item.type=="query") Vue.set(ref.queryresults, item.name, [].concat(result));
+                          if((item.type=="query")||(item.type=="enum")) Vue.set(ref.queryresults, item.name, [].concat(result));
                           // table is special.  if external data is passed.  we take that instead of results.
                           if(item.type=="table" && !ref.defaults(item.name)){
                             Vue.set(ref.form, item.name, [].concat(result));
@@ -1287,7 +1268,7 @@
                             if(restresult.status=="success"){
                               // console.log("expression for "+item.name+" triggered : result found -> "+ JSON.stringify(restresult.data.output));
                               if(item.type=="expression") Vue.set(ref.form, item.name, restresult.data.output);
-                              if(item.type=="query"){
+                              if((item.type=="query")||(item.type=="enum")){
                                 if(restresult.data.output==undefined){
                                   Vue.set(ref.queryresults, item.name, [])
                                 }else{
@@ -1361,7 +1342,7 @@
                         }
                         if(restresult.status=="success"){
                            //console.log("query "+item.name+" triggered : items found -> "+ restresult.data.output.length);
-                          if(item.type=="query"){
+                          if((item.type=="query")||(item.type=="enum")){
                             Vue.set(ref.queryresults, item.name, restresult.data.output);
                           }else{
                             Vue.set(ref.form, item.name, restresult.data.output);
@@ -1407,7 +1388,7 @@
               }else{  // not visible field
                 if(item.type=="expression"){
                   ref.defaultField(item.name)
-                }else if(item.type=="query" || item.type=="table"){
+                }else if(item.type=="query" || item.type=="enum" || item.type=="table"){
                   // console.log("resetting " + item.name)
                   ref.resetField(item.name)
                 }
@@ -1814,7 +1795,7 @@
             }
             Vue.set(ref.externalData,item.name,queryValue)
           }
-          if(["expression","query","table"].includes(item.type)){
+          if(["expression","query","enum","table"].includes(item.type)){
             Vue.set(ref.fieldOptions,item.name,{})                                // storing some easy to find options
             Vue.set(ref.fieldOptions[item.name],"valueColumn",item.valueColumn||"")
             Vue.set(ref.fieldOptions[item.name],"placeholderColumn",item.placeholderColumn||"")