+++ server/src/lib/logger.js
@@ -1,4 +1,5 @@
 const winston = require('winston');
+require('winston-daily-rotate-file');
 require('winston-syslog').Syslog;
 const loggerConfig = require('../../config/log.config');
 
@@ -23,22 +24,26 @@ const formatNoColor = winston.format.combine(
   ),
 )
 
-
-
 var transports = [
   new winston.transports.Console({
     stderrLevels: ["error"],
     level:loggerConfig.consolelevel,
     format:formatColor
   }),
-  new winston.transports.File({
-    filename: loggerConfig.path + "/ansibleforms.errors.log",
+  new winston.transports.DailyRotateFile({
+    filename: loggerConfig.path + "/ansibleforms.errors.%DATE%.log",
+    datePattern: 'YYYY-MM-DD',
+    maxFiles: '30d',
+    zippedArchive: true,
     level: 'error',
     format:formatNoColor
   }),
-  new winston.transports.File({
+  new winston.transports.DailyRotateFile({
     level: loggerConfig.level,
-    filename: loggerConfig.path + "/ansibleforms.log",
+    filename: loggerConfig.path + "/ansibleforms.%DATE%.log",
+    datePattern: 'YYYY-MM-DD',
+    zippedArchive: true,
+    maxFiles: '30d',    
     format:formatColor
   }),
 ]